+++ src/Support/Slim/ImagecacheRegister.php
@@ -1,55 +1,56 @@
 <?php namespace Onigoetz\Imagecache\Support\Slim;
 
+use GuzzleHttp\Psr7\LazyOpenStream;
 use Onigoetz\Imagecache\Exceptions\InvalidPresetException;
 use Onigoetz\Imagecache\Exceptions\NotFoundException;
 use Onigoetz\Imagecache\Manager;
 use Onigoetz\Imagecache\Transfer;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use RuntimeException;
-use Slim\Slim;
+use Slim\App;
 
 class ImagecacheRegister
 {
-    public static function register(Slim $app, $config)
+    public function request()
     {
-        $app->container->singleton(
-            'imagecache',
-            function () use ($config) {
-                return new Manager($config);
-            }
-        );
+        return function (ServerRequestInterface $req, ResponseInterface $res, $args) {
 
-        $url = "/{$config['path_web']}/{$config['path_cache']}/:preset/:file";
+            $preset = $args['preset'];
+            $file = $args['file'];
+
+            try {
+                $final_file = $this->imagecache->handleRequest($preset, $file);
+            } catch (InvalidPresetException $e) {
+                $res->getBody()->write($e->getMessage());
+                return $res->withStatus(404);
+            } catch (NotFoundException $e) {
+                $res->getBody()->write($e->getMessage());
+                return $res->withStatus(404);
+            } catch (RuntimeException $e) {
+                $res->getBody()->write($e->getMessage());
+                return $res->withStatus(500);
+            }
+
+            $transfer = new Transfer($final_file);
+            foreach ($transfer->getHeaders() as $key => $value) {
+                $res = $res->withHeader($key, $value);
+            }
+
+            return $res->withStatus($transfer->getStatus())->withBody(new LazyOpenStream($final_file, 'r'));
+        };
+    }
+
+    public static function register(App $app, $config)
+    {
+        $app->getContainer()['imagecache'] = function () use ($config) {
+            return new Manager($config);
+        };
 
         $app->get(
-            $url,
-            function ($preset, $file) use ($app) {
-                try {
-                    $final_file = $app->imagecache->handleRequest($preset, $file);
-                } catch (InvalidPresetException $e) {
-                    $app->response->setStatus(404);
-                    $app->response->body($e->getMessage());
-
-                    return;
-                } catch (NotFoundException $e) {
-                    $app->response->setStatus(404);
-                    $app->response->body($e->getMessage());
-
-                    return;
-                } catch (RuntimeException $e) {
-                    $app->response->setStatus(500);
-                    $app->response->body($e->getMessage());
-
-                    return;
-                }
-
-                $transfer = new Transfer($final_file);
-                $app->response->setStatus($transfer->getStatus());
-                $app->response->headers->replace($transfer->getHeaders());
-
-                $app->response->setBody(file_get_contents($final_file));
-            }
+            "/{$config['path_web']}/{$config['path_cache']}/{preset}/{file:.*}",
+            (new ImagecacheRegister)->request()
         )
-            ->conditions(['file' => '.*'])
-            ->name('onigoetz.imagecache');
+            ->setName('onigoetz.imagecache');
     }
 }